Split SO3 and O3; add reflection test for O3
